+++ src/aux-server/aux-web/aux-player/scene/PlayerSimulation3D.ts
@@ -448,7 +448,7 @@ export class PlayerSimulation3D extends Simulation3D {
                             this._pannable = calculateBooleanTagValue(
                                 calc,
                                 bot,
-                                `auxContextPannable`,
+                                `auxContextPlayerPannable`,
                                 true
                             );