+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     -   Setting `auxPositioningMode` to `absolute` will no longer prevent mods.
     -   Changed `applyMod()` and `subtractMods()` to not send `onMod()` events.
     -   Renamed the `diffs` property on the `onModDrop` argument to `mod`.
+    -   Renamed the following tags:
+        -   `auxContextPannable` -> `auxContextPlayerPannable`
 -   Improvements
     -   Added `onModDropEnter` and `onModDropExit` listeners for when a mod is dragged onto or off of a bot.
         -   The bot that the mod will be applied to recieves the `onModDropEnter` and `onModDropExit` events.